+++ src/ionic-build.ts
@@ -5,6 +5,7 @@ import * as path from 'path';
 import { Project } from './project';
 import { MonoRepoType } from './monorepo';
 import { ionicState } from './ionic-tree-provider';
+import { InternalCommand } from './command-name';
 
 /**
  * Creates the ionic build command
@@ -13,7 +14,7 @@ import { ionicState } from './ionic-tree-provider';
  */
 export function ionicBuild(project: Project, configurationArg?: string): string {
 	// For convenience, check if npm install was done and do it
-	const nmf = path.join(project.folder, 'node_modules');
+	const nmf =project.getNodeModulesFolder();
 	const preop = (!fs.existsSync(nmf)) ? 'npm install && ' : '';
 
 	const prod: boolean = vscode.workspace.getConfiguration('ionic').get('buildForProduction');	
@@ -21,6 +22,7 @@ export function ionicBuild(project: Project, configurationArg?: string): string 
 		case MonoRepoType.none: return `${preop}${ionicCLIBuild(prod, configurationArg)}`;
 		case MonoRepoType.npm: return `${preop}${ionicCLIBuild(prod, configurationArg)} --project=${ionicState.workspace}`;
 		case MonoRepoType.nx: return `${preop}${nxBuild(prod, project)}`;
+		case MonoRepoType.folder: return `${InternalCommand.cwd}${preop}${ionicCLIBuild(prod, configurationArg)}`;
 		default: throw new Error('Unsupported Monorepo type');
 	}
 }